requirements.txt
aerich==0.8.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
asyncclick==8.1.8.0
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dictdiffer==0.9.0
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pypika-tortoise==0.5.0
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
tomli_w==1.2.0
tomlkit==0.13.2
tortoise-orm==0.24.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
